requirements.txt
keras
numpy
tensorflow
progressbar2